added the react router
