+++ app/javascript/react/src/index.js
@@ -1,1 +1,1 @@
-import Home from "./components/Home.jsx";
+import App from "./components/App.jsx";
