+++ package.json
@@ -9,6 +9,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^9.1.0",
+    "react-router-dom": "^6.21.2",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   },
